src/replace_pronouns/api.py
import os
import openai
from dotenv import load_dotenv
from typing import List
from typing import Union
import requests

load_dotenv()  # read local .env file
geminipro_api_key = os.getenv("GEMINIPRO_API_KEY")
geminipro_model = "gemini-1.5-pro"
geminipro_url = os.getenv("GEMINIPRO_API_BASE_URL")

client = openai.OpenAI(api_key=geminipro_api_key,base_url=geminipro_url)

def get_completion(
    prompt: str,
    system_message: str = "You are a helpful assistant.",
    model: str = geminipro_model,
    temperature: float = 0.3,
    json_mode: bool = False,
) -> Union[str, dict]:
    """
        Generate a completion using the OpenAI API.

    Args:
        prompt (str): The user's prompt or query.
        system_message (str, optional): The system message to set the context for the assistant.
            Defaults to "You are a helpful assistant.".
        model (str, optional): The name of the OpenAI model to use for generating the completion.
            Defaults to "gpt-4-turbo".
        temperature (float, optional): The sampling temperature for controlling the randomness of the generated text.
            Defaults to 0.3.
        json_mode (bool, optional): Whether to return the response in JSON format.
            Defaults to False.

    Returns:
        Union[str, dict]: The generated completion.
            If json_mode is True, returns the complete API response as a dictionary.
            If json_mode is False, returns the generated text as a string.
    """

    if json_mode:
        response = client.chat.completions.create(
            model=model,
            temperature=temperature,
            top_p=1,
            response_format={"type": "json_object"},
            messages=[
                {"role": "system", "content": system_message},
                {"role": "user", "content": prompt},
            ],
        )
        return response.choices[0].message.content
    else:
        response = client.chat.completions.create(
            model=model,
            temperature=temperature,
            top_p=1,
            messages=[
                {"role": "system", "content": system_message},
                {"role": "user", "content": prompt},
            ],
        )
        return response.choices[0].message.content
